requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
Flask-Script==2.0.6
flask_paginate==2022.1.8
Flask-WTF==1.1.1
furl==2.1.2
oauthlib==3.2.1
python-dotenv==0.21.0
requests==2.28.1
